Move exprtrace package to internal/exprtrace
