+++ internal/exprtrace/tree_printer.go
